+++ SerialImplmentation/TravelComputer/arrayManipulation.h
@@ -1,5 +1,5 @@
 #include<stdio.h>
 #include <stdlib.h>
 
-int* remove_element(int* arr, int size, int pos);
+short* remove_element(short* arr, short size, short pos);
 short* add_element(short* arr, short size, short element);
